+++ src/serial_chain_robots/serial_chain_robot_dynamics.cpp
@@ -1,8 +1,8 @@
 # include <iostream>
 # include <Eigen/Dense>
 
-# include "multi_joint_robots/serial_chain_robot_helper_funs.h"
-# include "multi_joint_robots/serial_chain_robot_dynamics.h"
+# include "serial_chain_robots/serial_chain_robot_helper_funs.h"
+# include "serial_chain_robots/serial_chain_robot_dynamics.h"
 
 using Eigen::MatrixXd;
 using Eigen::VectorXd;
