+++ src/serial_chain_robots/serial_chain_robot_helper_funs.h
