+++ test/serial_chain_robots/build_test_serial_chain_robot_helper_funs.sh
@@ -18,8 +18,8 @@ LOCAL_INCLUDE_PATH="/Users/priyankabhovad/Downloads/cpp_ex/system_modeling_and_c
 # Include flags for local header files
 LOCAL_INCLUDE_FLAGS="-I$LOCAL_INCLUDE_PATH"
 
-SRCS="test/multi_joint_robots/test_serial_chain_robot_helper_funs.cpp 
-     src/multi_joint_robots/serial_chain_robot_helper_funs.cpp"
+SRCS="test/serial_chain_robots/test_serial_chain_robot_helper_funs.cpp 
+     src/serial_chain_robots/serial_chain_robot_helper_funs.cpp"
 
 # Output executable name
 TARGET="build/myprogram"
