+++ test/serial_chain_robots/test_forward_kinematics.cpp
@@ -3,7 +3,7 @@
 # include <cmath>
 # include <Eigen/Dense>
 
-# include "multi_joint_robots/forward_kinematics.h"
+# include "serial_chain_robots/forward_kinematics.h"
 using Eigen::MatrixXd;
 using Eigen::VectorXd;
 using Eigen::Matrix3d;
