+++ test/serial_chain_robots/test_serial_chain_robot_dynamics.cpp
@@ -3,7 +3,7 @@
 # include <cmath>
 # include <Eigen/Dense>
 
-# include "multi_joint_robots/serial_chain_robot_dynamics.h"
+# include "serial_chain_robots/serial_chain_robot_dynamics.h"
 
 using Eigen::MatrixXd;
 using Eigen::VectorXd;
